Run a c file through preprocessor and save to file
From the man page for `gcc`:
option `-E` stops the compilation process after the preprocessing stage
and does not run the compiler proper.  The output is in the form of
preprocessed source code, which is sent to the standard output.

option `-o file` places output in file `file`. This applies to whatever
sort of output is being produced, whether it be an executable file, an
object file, an assembler file or preprocessed C code.