Compile c file but do not link
From the man page for gcc:
The `-c` option stops the compilation process at the compile stage
and does not link.  The linking stage simply is not done. The ultimate
output is in the form of an object file for each source file.
By default, the object file name for a source file is made by
replacing the suffix `.c`, `.i`, `.s`, etc., with `.o`.