Generate assembly code for a C file
From the man page for gcc:
The `-S` option stops after the stage of compilation proper and does not
assemble. The output is in the form of an assembler code file for each
non-assembler input file specified. By default, the assembler file name
for a source file is made by replacing the suffix `.c`, `.i` etc
with `.s`.